nfs4/server: fix signed-char sign extension in proxy_stateid_extract_boot_seq
stateid4.other is char[12] (rpcgen translates XDR opaque to char,
not unsigned char).  C leaves char's signedness implementation-
defined: signed on x86/x86_64, unsigned on aarch64 and most others.

proxy_stateid_extract_boot_seq did

    return ((uint16_t)stid->other[0] << 8) |
           (uint16_t)stid->other[1];

On a signed-char platform, a byte with the high bit set (>= 0x80)
sign-extends through integer promotion *before* the (uint16_t) cast.
A stateid encoded with low byte 0xFF decodes to 0xFFFF instead of
0x00FF, so any boot_seq with a high-bit byte round-trips wrong on
x86_64.  Reffs-ci-hetzner (Fedora x86_64) hits this; dreamer
(Fedora aarch64, char unsigned) does not.

Cast through uint8_t before the uint16_t conversion so the byte
is zero-extended on every platform:

    (uint16_t)(uint8_t)stid->other[0]

The fix is a no-op on platforms where char is already unsigned and
a correctness fix everywhere else.

The matching test_alloc_boot_seq_big_endian had the same bug --
ck_assert_uint_eq(s.other[1], 0xAB) on signed char promoted to
int(-85), then to uint64_t (huge).  Cast through uint8_t there too.

test_extract_round_trip with seqs[i] = 0x00FF was the canary: it
asserted extract returns 0x00FF, got 0xFFFF, and the test pinned
the round-trip exactly.

+++ lib/nfs4/server/proxy_stateid.c
@@ -103,8 +103,19 @@ uint16_t proxy_stateid_extract_boot_seq(const stateid4 *stid)
 {
 	if (!stid)
 		return 0;
-	return ((uint16_t)stid->other[PROXY_STATEID_BOOT_SEQ_OFF + 0] << 8) |
-	       (uint16_t)stid->other[PROXY_STATEID_BOOT_SEQ_OFF + 1];
+	/*
+	 * stateid4.other is char[12].  C leaves char's signedness up to
+	 * the implementation: signed on x86/x86_64, unsigned on aarch64
+	 * and many others.  Without the (uint8_t) cast, a byte with the
+	 * high bit set (>= 0x80) sign-extends through integer promotion
+	 * before the (uint16_t) cast, so e.g. byte 0xAB becomes 0xFFAB
+	 * instead of 0x00AB.  The static_asserts above guarantee the
+	 * offsets are correct; this guarantees the bytes themselves
+	 * decode portably.
+	 */
+	return ((uint16_t)(uint8_t)stid->other[PROXY_STATEID_BOOT_SEQ_OFF + 0]
+		<< 8) |
+	       (uint16_t)(uint8_t)stid->other[PROXY_STATEID_BOOT_SEQ_OFF + 1];
 }
 
 bool proxy_stateid_is_stale(const stateid4 *stid, uint16_t current_boot_seq)
